test: file the tests inherited from main under tests/unit
The rebase brought in main's new test modules at the old flat layout —
four at the root and a `tests/monitoring` package — next to the
unit/integration split this branch introduced. They now live under
`tests/unit` (monitoring as a subpackage) and follow the same shape as the
rest: one test class per module. `tests/support.py` stays beside
`conftest.py` as the shared helper it is, and the OpenAPI snapshot stays
where the SDK contract test reads it from.
